requirements.txt
djangorestframework
PyMySQL